+++ core/src/main/java/org/svip/sbomfactory/generators/parsers/packagemanagers/GradleParser.java
@@ -24,7 +24,13 @@ import static org.svip.sbomfactory.generators.utils.Debug.*;
 public class GradleParser extends PackageManagerParser {
     //#region Constructors
 
-    public GradleParser() { super("https://docs.gradle.org/current/javadoc/", new JsonFactory()); }
+    public GradleParser() {
+        super(
+                "https://docs.gradle.org/current/javadoc/",
+                new JsonFactory(),
+                "\\$([^\\n/\\\\]*)" // Regex101: https://regex101.com/r/1Y2gb5/2
+        );
+    }
 
     //#endregion
 
@@ -32,95 +38,34 @@ public class GradleParser extends PackageManagerParser {
 
     @Override
     protected void parseData(ArrayList<ParserComponent> components, HashMap<String, Object> data) {
-        // Init dependencies list
-        this.dependencies = new ArrayList<>();
-
-        // Get raw data
-        final String rawDependencies = data.get("dependencies").toString();
-
-        // Init dependency Pattern
-        // Regex101: https://regex101.com/r/cFnCpF/12
-        final Pattern dependenciesPattern =
-                Pattern.compile("^\\s*\\w* ?\\(?\\[?(?:(?:(?:[\\\"'](.*:.*)[\\\"'])|(?:(group: [^\\n{]*, name: [^\\n{]*, version: [^\\n{)\\]]*)\\)?\\[?))|file.*\\((.*)?\\))(?:(?=.*\\{\\n?).*\\{([\\S\\s]*?)^\\t?(?: {4})?\\})?", Pattern.MULTILINE);
-
-        // Init dependency matcher
-        final Matcher depMatcher = dependenciesPattern.matcher(rawDependencies);
-
-        // Iterate over results
-        for (final MatchResult depResult : depMatcher.results().toList()) {
-            final LinkedHashMap<String, String> dep = new LinkedHashMap<>();
-            // Group 1 format: "org.springframework:spring-api:3.6"
-            if(depResult.group(1) != null) {
-                // Extract values from dep string
-                final String[] depValues = depResult.group(1).split(":");
-                dep.put("groupId", depValues[0]);
-                if(depValues.length > 1) dep.put("artifactId", depValues[1]);
-                if(depValues.length > 2) dep.put("version", depValues[2]);
-            } // Group 2 format: "group: 'org.springframework', name: 'spring-core', version: '2.5'"
-            else if(depResult.group(2) != null) {
-                final String[] depValues = depResult.group(2).split(",");
-                final HashMap<String, String> props = new HashMap<>(depValues.length);
-                Arrays.stream(depValues).forEach(d -> {
-                    d = d.replace("'", "").replace("\"", "");
-                    final String dKey = d.substring(0, d.indexOf(":")).trim();
-                    final String dValue = d.substring(d.indexOf(":") + 1).trim();
-                    props.put(dKey, dValue);
-                });
-                if(props.containsKey("group")) dep.put("groupId", props.get("group"));
-                if(props.containsKey("name")) dep.put("artifactId", props.get("name"));
-                if(props.containsKey("version")) dep.put("version", props.get("version"));
-            } // Group 3 format for file(s)/dir(s): "'hibernate.jar', 'libs/spring.jar'"
-            else if(depResult.group(3) != null) {
-                // Extract values from dep string
-                final String[] depValues =  depResult.group(3)
-                        .replace("'", "")
-                        .replace("\"", "")
-                        .split(",");
-
-                // If one file is found, add its data to dep
-                if (depValues.length == 1) dep.put("artifactId", depValues[0].trim());
-                // If more than one file is found, create and add a new LHM for each one
-                else {
-                    this.dependencies.add(new LinkedHashMap<>() {{
-                        put("artifactId", depValues[0].trim());
-                    }});
-                    continue;
-                }
-            }
-
-            // Add any values found in group 4, if present
-            // This can only occur when group 1 or 2 != null
-            if(depResult.group(4) != null) {
-                // TODO: Process group 4 values
-                //  Group 4 captures
-            }
-
-            // Insert value
-            this.dependencies.add(dep); // TODO: Uniqueness? Tests fail on non-unique artifactIds
-        }
+        // Init properties
+        this.properties = new HashMap<>();
 
         // Get properties
         final ArrayList<String> ext =
                 (ArrayList<String>) data.get("ext");
 
         // Store properties
-        this.properties = (HashMap<String, String>) ext
+        this.resolveProperties((HashMap<String, String>) ext
                 .stream().collect(
                         Collectors.toMap(
                                 e -> e.substring(0, e.indexOf('=')).trim(),
-                                e -> e.substring(e.indexOf('=') + 1).trim())
-                );
+                                e -> e.substring(e.indexOf('=') + 1)
+                                        .trim()
+                                        .replace("'", "")
+                                        .replace("\"", "")
+                        )));
 
         // Iterate over dependencies
         for (final LinkedHashMap<String, String> d : this.dependencies) {
             // Create ParserComponent from dep info
             final ParserComponent c = new ParserComponent(d.get("artifactId"));
             c.setGroup(d.get("group"));
-            if(d.containsKey("type")) {
+            if (d.containsKey("type")) {
                 final String type = d.get("type");
-                if(type.equals("file")) c.setType(ParserComponent.Type.INTERNAL);
+                if (type.equals("file")) c.setType(ParserComponent.Type.INTERNAL);
             }
-            if(d.containsKey("version")) c.setVersion(d.get("version"));
+            if (d.containsKey("version")) c.setVersion(d.get("version"));
 
             // TODO: Query
 //            String url = "";
@@ -138,24 +83,17 @@ public class GradleParser extends PackageManagerParser {
     }
 
     @Override
-    protected void resolveProperties(HashMap<String, String> props) {
-
-    }
-
-    @Override
-    protected void resolveProperty(String key, String value, HashMap<String, String> props, Pattern p) {
-
-    }
-
-    @Override
     public void parse(ArrayList<ParserComponent> components, String fileContents) {
         // Init main data structure
         final LinkedHashMap<String, Object> data = new LinkedHashMap<>();
 
+        // Init this.dependencies
+        this.dependencies = new ArrayList<>();
+
         // Init main Matcher
         // Regex101: https://regex101.com/r/a3rIlp/3
         final Matcher m = Pattern.compile("^(.*) \\{([\\s\\S]*?)^\\}", Pattern.MULTILINE)
-                        .matcher(fileContents);
+                .matcher(fileContents);
 
         // Store results in data
         for (final MatchResult mr : m.results().toList()) {
@@ -164,7 +102,73 @@ public class GradleParser extends PackageManagerParser {
             final String value = mr.group(2).trim().replace("\r", "");
 
             // Dependencies will need to be parsed further, so pass raw string
-            if (key.equals("dependencies")) data.put(key, value);
+            if (key.equals("dependencies")) {
+                // Init dependencies list
+                final ArrayList<LinkedHashMap<String, String>> deps = new ArrayList<>();
+
+                // Init dependency Pattern
+                // Regex101: https://regex101.com/r/cFnCpF/12
+                final Pattern dependenciesPattern =
+                        Pattern.compile("^\\s*\\w* ?\\(?\\[?(?:(?:(?:[\\\"'](.*:.*)[\\\"'])|(?:(group: [^\\n{]*, name: [^\\n{]*, version: [^\\n{)\\]]*)\\)?\\[?))|file.*\\((.*)?\\))(?:(?=.*\\{\\n?).*\\{([\\S\\s]*?)^\\t?(?: {4})?\\})?", Pattern.MULTILINE);
+
+                // Init dependency matcher
+                final Matcher depMatcher = dependenciesPattern.matcher(value);
+
+                // Iterate over results
+                for (final MatchResult depResult : depMatcher.results().toList()) {
+                    final LinkedHashMap<String, String> dep = new LinkedHashMap<>();
+                    // Group 1 format: "org.springframework:spring-api:3.6"
+                    if(depResult.group(1) != null) {
+                        // Extract values from dep string
+                        final String[] depValues = depResult.group(1).split(":");
+                        dep.put("groupId", depValues[0]);
+                        if(depValues.length > 1) dep.put("artifactId", depValues[1]);
+                        if(depValues.length > 2) dep.put("version", depValues[2]);
+                    } // Group 2 format: "group: 'org.springframework', name: 'spring-core', version: '2.5'"
+                    else if(depResult.group(2) != null) {
+                        final String[] depValues = depResult.group(2).split(",");
+                        final HashMap<String, String> props = new HashMap<>(depValues.length);
+                        Arrays.stream(depValues).forEach(d -> {
+                            d = d.replace("'", "").replace("\"", "");
+                            final String dKey = d.substring(0, d.indexOf(":")).trim();
+                            final String dValue = d.substring(d.indexOf(":") + 1).trim();
+                            props.put(dKey, dValue);
+                        });
+                        if(props.containsKey("group")) dep.put("groupId", props.get("group"));
+                        if(props.containsKey("name")) dep.put("artifactId", props.get("name"));
+                        if(props.containsKey("version")) dep.put("version", props.get("version"));
+                    } // Group 3 format for file(s)/dir(s): "'hibernate.jar', 'libs/spring.jar'"
+                    else if(depResult.group(3) != null) {
+                        // Extract values from dep string
+                        final String[] depValues =  depResult.group(3)
+                                .replace("'", "")
+                                .replace("\"", "")
+                                .split(",");
+
+                        // If one file is found, add its data to dep
+                        if (depValues.length == 1) dep.put("artifactId", depValues[0].trim());
+                            // If more than one file is found, create and add a new LHM for each one
+                        else {
+                            deps.add(new LinkedHashMap<>() {{
+                                put("artifactId", depValues[0].trim());
+                            }});
+                            continue;
+                        }
+                    }
+
+                    // Add any values found in group 4, if present
+                    // This can only occur when group 1 or 2 != null
+                    if(depResult.group(4) != null) {
+                        // TODO: Process group 4 values
+                        //  Group 4 captures
+                    }
+
+                    // Insert value
+                    deps.add(dep); // TODO: Uniqueness? Tests fail on non-unique artifactIds
+                }
+                // Insert data
+                this.dependencies.addAll(deps);
+            }
             // Other collected data can be split on "\n"
             else data.put(key, new ArrayList<>(Arrays.stream(value.split("\n")).map(String::trim).toList()));
         }